fix: continue refactor helpers
