+++ packages/spec/supply/setCollateral.spec.ts
@@ -13,7 +13,7 @@ import {
 } from '@aave/client-next/test-utils';
 import { sendWith } from '@aave/client-next/viem';
 import { beforeAll, describe, expect, it } from 'vitest';
-import { findReserveToSupply, supplyToReserve } from '../borrow/helper';
+import { findReserveToSupply, supplyToReserve } from '../helpers/borrowSupply';
 import { assertNonEmptyArray, assertSingleElementArray } from '../test-utils';
 
 const user = await createNewWallet();
